Move payload_dict/merge_dps_results out of Device, remove duplicated functions
