+++ tinytuya/core/XenonDevice.py
@@ -16,6 +16,7 @@ from .error_helper import ERR_CONNECT, ERR_DEVTYPE, ERR_JSON, ERR_KEY_OR_VER, ER
 from .exceptions import DecodeError
 from .message_helper import MessagePayload, TuyaMessage, pack_message, unpack_message, parse_header
 from . import command_types as CT, header as H
+from .core import merge_dps_results
 
 log = logging.getLogger(__name__)
 
@@ -83,129 +84,6 @@ def device_info( dev_id ):
         pass
 
     return devinfo
-
-def merge_dps_results(dest, src):
-    """Merge multiple receive() responses into a single dict
-
-    `src` will be combined with and merged into `dest`
-    """
-    if src and isinstance(src, dict) and 'Error' not in src and 'Err' not in src:
-        for k in src:
-            if k == 'dps' and src[k] and isinstance(src[k], dict):
-                if 'dps' not in dest or not isinstance(dest['dps'], dict):
-                    dest['dps'] = {}
-                for dkey in src[k]:
-                    dest['dps'][dkey] = src[k][dkey]
-            elif k == 'data' and src[k] and isinstance(src[k], dict) and 'dps' in src[k] and isinstance(src[k]['dps'], dict):
-                if k not in dest or not isinstance(dest[k], dict):
-                    dest[k] = {'dps': {}}
-                if 'dps' not in dest[k] or not isinstance(dest[k]['dps'], dict):
-                    dest[k]['dps'] = {}
-                for dkey in src[k]['dps']:
-                    dest[k]['dps'][dkey] = src[k]['dps'][dkey]
-            else:
-                dest[k] = src[k]
-
-# Tuya Device Dictionary - Command and Payload Overrides
-#
-# 'default' devices require the 0a command for the DP_QUERY request
-# 'device22' devices require the 0d command for the DP_QUERY request and a list of
-#            dps used set to Null in the request payload
-#
-# Any command not defined in payload_dict will be sent as-is with a
-#  payload of {"gwId": "", "devId": "", "uid": "", "t": ""}
-
-payload_dict = {
-    # Default Device
-    "default": {
-        CT.AP_CONFIG: {  # [BETA] Set Control Values on Device
-            "command": {"gwId": "", "devId": "", "uid": "", "t": ""},
-        },
-        CT.CONTROL: {  # Set Control Values on Device
-            "command": {"devId": "", "uid": "", "t": ""},
-        },
-        CT.STATUS: {  # Get Status from Device
-            "command": {"gwId": "", "devId": ""},
-        },
-        CT.HEART_BEAT: {"command": {"gwId": "", "devId": ""}},
-        CT.DP_QUERY: {  # Get Data Points from Device
-            "command": {"gwId": "", "devId": "", "uid": "", "t": ""},
-        },
-        CT.CONTROL_NEW: {"command": {"devId": "", "uid": "", "t": ""}},
-        CT.DP_QUERY_NEW: {"command": {"devId": "", "uid": "", "t": ""}},
-        CT.UPDATEDPS: {"command": {"dpId": [18, 19, 20]}},
-        CT.LAN_EXT_STREAM: { "command": { "reqType": "", "data": {} }},
-    },
-    # Special Case Device with 22 character ID - Some of these devices
-    # Require the 0d command as the DP_QUERY status request and the list of
-    # dps requested payload
-    "device22": {
-        CT.DP_QUERY: {  # Get Data Points from Device
-            "command_override": CT.CONTROL_NEW,  # Uses CONTROL_NEW command for some reason
-            "command": {"devId": "", "uid": "", "t": ""},
-        },
-    },
-    # v3.3+ devices do not need devId/gwId/uid
-    "v3.4": {
-        CT.CONTROL: {
-            "command_override": CT.CONTROL_NEW,  # Uses CONTROL_NEW command
-            "command": {"protocol":5, "t": "int", "data": {}}
-            },
-        CT.CONTROL_NEW: {
-            "command": {"protocol":5, "t": "int", "data": {}}
-        },
-        CT.DP_QUERY: {
-            "command_override": CT.DP_QUERY_NEW,
-            "command": {} #"protocol":4, "t": "int", "data": {}}
-        },
-        CT.DP_QUERY_NEW: {
-            "command": {}
-        },
-    },
-    # v3.5 is just a copy of v3.4
-    "v3.5": {
-        CT.CONTROL: {
-            "command_override": CT.CONTROL_NEW,  # Uses CONTROL_NEW command
-            "command": {"protocol":5, "t": "int", "data": {}}
-        },
-        CT.CONTROL_NEW: {
-            "command": {"protocol":5, "t": "int", "data": {}}
-        },
-        CT.DP_QUERY: {
-            "command_override": CT.DP_QUERY_NEW,
-            "command": {}
-        },
-        CT.DP_QUERY_NEW: {
-            "command": {}
-        },
-    },
-    # placeholders, not yet needed
-    "gateway": { },
-    "gateway_v3.4": { },
-    "gateway_v3.5": { },
-    "zigbee": {
-        CT.CONTROL: { "command": {"t": "int", "cid": ""} },
-        CT.DP_QUERY: { "command": {"t": "int", "cid": ""} },
-    },
-    "zigbee_v3.4": {
-        CT.CONTROL: {
-            "command_override": CT.CONTROL_NEW,
-            "command": {"protocol":5, "t": "int", "data": {"cid":""}}
-        },
-        CT.CONTROL_NEW: {
-            "command": {"protocol":5, "t": "int", "data": {"cid":""}}
-        },
-    },
-    "zigbee_v3.5": {
-        CT.CONTROL: {
-            "command_override": CT.CONTROL_NEW,
-            "command": {"protocol":5, "t": "int", "data": {"cid":""}}
-        },
-        CT.CONTROL_NEW: {
-            "command": {"protocol":5, "t": "int", "data": {"cid":""}}
-        },
-    },
-}
 
 class XenonDevice(object):
     def __init__(
@@ -1200,7 +1078,7 @@ class XenonDevice(object):
             return result
 
         # dict2 will be merged into dict1
-        # as dict2 is payload_dict['...'] we only need to worry about copying 2 levels deep,
+        # as dict2 is CT.payload_dict['...'] we only need to worry about copying 2 levels deep,
         #  the command id and "command"/"command_override" keys: i.e. dict2[CMD_ID]["command"]
         def _merge_payload_dicts(dict1, dict2):
             for cmd in dict2:
@@ -1222,19 +1100,19 @@ class XenonDevice(object):
         #   'zigbee_'+[version string] if sub-device - [dev_type if not "default"]
         if not self.payload_dict or self.last_dev_type != self.dev_type:
             self.payload_dict = {}
-            _merge_payload_dicts( self.payload_dict, payload_dict['default'] )
+            _merge_payload_dicts( self.payload_dict, CT.payload_dict['default'] )
             if self.children:
-                _merge_payload_dicts( self.payload_dict, payload_dict['gateway'] )
+                _merge_payload_dicts( self.payload_dict, CT.payload_dict['gateway'] )
             if self.cid:
-                _merge_payload_dicts( self.payload_dict, payload_dict['zigbee'] )
-            if self.version_str in payload_dict:
-                _merge_payload_dicts( self.payload_dict, payload_dict[self.version_str] )
-            if self.children and ('gateway_'+self.version_str) in payload_dict:
-                _merge_payload_dicts( self.payload_dict, payload_dict['gateway_'+self.version_str] )
-            if self.cid and ('zigbee_'+self.version_str) in payload_dict:
-                _merge_payload_dicts( self.payload_dict, payload_dict['zigbee_'+self.version_str] )
+                _merge_payload_dicts( self.payload_dict, CT.payload_dict['zigbee'] )
+            if self.version_str in CT.payload_dict:
+                _merge_payload_dicts( self.payload_dict, CT.payload_dict[self.version_str] )
+            if self.children and ('gateway_'+self.version_str) in CT.payload_dict:
+                _merge_payload_dicts( self.payload_dict, CT.payload_dict['gateway_'+self.version_str] )
+            if self.cid and ('zigbee_'+self.version_str) in CT.payload_dict:
+                _merge_payload_dicts( self.payload_dict, CT.payload_dict['zigbee_'+self.version_str] )
             if self.dev_type != 'default':
-                _merge_payload_dicts( self.payload_dict, payload_dict[self.dev_type] )
+                _merge_payload_dicts( self.payload_dict, CT.payload_dict[self.dev_type] )
             log.debug( 'final payload_dict for %r (%r/%r): %r', self.id, self.version_str, self.dev_type, self.payload_dict )
             # save it so we don't have to calculate this again unless something changes
             self.last_dev_type = self.dev_type
