+++ tinytuya/core/command_types.py
@@ -26,3 +26,105 @@ AP_CONFIG_NEW   = 0x14 # 20 # FRM_AP_CFG_WF_V40
 BOARDCAST_LPV34 = 0x23 # 35 # FR_TYPE_BOARDCAST_LPV34
 REQ_DEVINFO     = 0x25 # broadcast to port 7000 to get v3.5 devices to send their info
 LAN_EXT_STREAM  = 0x40 # 64 # FRM_LAN_EXT_STREAM
+
+
+# Tuya Device Dictionary - Command and Payload Overrides
+#
+# 'default' devices require the 0a command for the DP_QUERY request
+# 'device22' devices require the 0d command for the DP_QUERY request and a list of
+#            dps used set to Null in the request payload
+#
+# Any command not defined in payload_dict will be sent as-is with a
+#  payload of {"gwId": "", "devId": "", "uid": "", "t": ""}
+
+payload_dict = {
+    # Default Device
+    "default": {
+        AP_CONFIG: {  # [BETA] Set Control Values on Device
+            "command": {"gwId": "", "devId": "", "uid": "", "t": ""},
+        },
+        CONTROL: {  # Set Control Values on Device
+            "command": {"devId": "", "uid": "", "t": ""},
+        },
+        STATUS: {  # Get Status from Device
+            "command": {"gwId": "", "devId": ""},
+        },
+        HEART_BEAT: {"command": {"gwId": "", "devId": ""}},
+        DP_QUERY: {  # Get Data Points from Device
+            "command": {"gwId": "", "devId": "", "uid": "", "t": ""},
+        },
+        CONTROL_NEW: {"command": {"devId": "", "uid": "", "t": ""}},
+        DP_QUERY_NEW: {"command": {"devId": "", "uid": "", "t": ""}},
+        UPDATEDPS: {"command": {"dpId": [18, 19, 20]}},
+        LAN_EXT_STREAM: { "command": { "reqType": "", "data": {} }},
+    },
+    # Special Case Device with 22 character ID - Some of these devices
+    # Require the 0d command as the DP_QUERY status request and the list of
+    # dps requested payload
+    "device22": {
+        DP_QUERY: {  # Get Data Points from Device
+            "command_override": CONTROL_NEW,  # Uses CONTROL_NEW command for some reason
+            "command": {"devId": "", "uid": "", "t": ""},
+        },
+    },
+    # v3.3+ devices do not need devId/gwId/uid
+    "v3.4": {
+        CONTROL: {
+            "command_override": CONTROL_NEW,  # Uses CONTROL_NEW command
+            "command": {"protocol":5, "t": "int", "data": {}}
+            },
+        CONTROL_NEW: {
+            "command": {"protocol":5, "t": "int", "data": {}}
+        },
+        DP_QUERY: {
+            "command_override": DP_QUERY_NEW,
+            "command": {} #"protocol":4, "t": "int", "data": {}}
+        },
+        DP_QUERY_NEW: {
+            "command": {}
+        },
+    },
+    # v3.5 is just a copy of v3.4
+    "v3.5": {
+        CONTROL: {
+            "command_override": CONTROL_NEW,  # Uses CONTROL_NEW command
+            "command": {"protocol":5, "t": "int", "data": {}}
+        },
+        CONTROL_NEW: {
+            "command": {"protocol":5, "t": "int", "data": {}}
+        },
+        DP_QUERY: {
+            "command_override": DP_QUERY_NEW,
+            "command": {}
+        },
+        DP_QUERY_NEW: {
+            "command": {}
+        },
+    },
+    # placeholders, not yet needed
+    "gateway": { },
+    "gateway_v3.4": { },
+    "gateway_v3.5": { },
+    "zigbee": {
+        CONTROL: { "command": {"t": "int", "cid": ""} },
+        DP_QUERY: { "command": {"t": "int", "cid": ""} },
+    },
+    "zigbee_v3.4": {
+        CONTROL: {
+            "command_override": CONTROL_NEW,
+            "command": {"protocol":5, "t": "int", "data": {"cid":""}}
+        },
+        CONTROL_NEW: {
+            "command": {"protocol":5, "t": "int", "data": {"cid":""}}
+        },
+    },
+    "zigbee_v3.5": {
+        CONTROL: {
+            "command_override": CONTROL_NEW,
+            "command": {"protocol":5, "t": "int", "data": {"cid":""}}
+        },
+        CONTROL_NEW: {
+            "command": {"protocol":5, "t": "int", "data": {"cid":""}}
+        },
+    },
+}
